JP-BE-PY-batch-1/django-app/project_1/requirements.txt
djangorestframework
faker
django-filter